Modify app padding
