+++ src/components/quests/QuestCallToAction/gavel.png
