+++ src/components/quests/QuestPageHeader/QuestPageHeader.tsx
@@ -1,77 +1,19 @@
-import type { PageHeaderProps } from "@/components/app";
-import {
-  Button,
-  IconText,
-  Link,
-  TimeAgo,
-  Tooltip,
-  TooltipTrigger,
-} from "@/components/common";
+import { PageHeader, type PageHeaderProps } from "@/components/app";
+import { Button, Tooltip, TooltipTrigger } from "@/components/common";
 import {
   EditQuestTitleModal,
-  QuestCategory,
-  QuestCosts,
-  QuestJurisdiction,
-  QuestTimeRequired,
+  QuestPageBadges,
+  QuestPageToolbar,
+  StatusSelect,
 } from "@/components/quests";
-import { useIsMobile } from "@/utils/useIsMobile";
 import { api } from "@convex/_generated/api";
 import type { Doc, Id } from "@convex/_generated/dataModel";
 import type { Status } from "@convex/constants";
 import { useNavigate } from "@tanstack/react-router";
-import { Authenticated, useMutation, useQuery } from "convex/react";
-import { ArrowLeft, Clock, Pencil } from "lucide-react";
+import { useMutation } from "convex/react";
+import { Pencil } from "lucide-react";
 import { useState } from "react";
-import { Heading } from "react-aria-components";
 import { toast } from "sonner";
-import { StatusSelect } from "../StatusSelect/StatusSelect";
-
-interface QuestPageToolbarProps {
-  quest?: Doc<"quests"> | null;
-  editable?: boolean;
-}
-
-function QuestPageToolbar({ quest, editable }: QuestPageToolbarProps) {
-  const user = useQuery(api.users.getCurrent);
-  const canEdit = user?.role === "admin";
-  const updatedTime = quest?.updatedAt ? (
-    <TimeAgo date={new Date(quest.updatedAt)} />
-  ) : (
-    "some time ago"
-  );
-
-  return (
-    <div className="flex items-center justify-between h-9 pb-1 w-full">
-      <div className="flex gap-2 items-center">
-        <IconText icon={Clock}>Last edited {updatedTime}</IconText>
-      </div>
-      <Authenticated>
-        {editable && (
-          <div className="flex gap-4 items-center">
-            <Link
-              button={{ size: "small", variant: "success" }}
-              href={{
-                search: { edit: undefined },
-              }}
-            >
-              Save changes
-            </Link>
-          </div>
-        )}
-        {!editable && canEdit && (
-          <Link
-            button={{ size: "small", variant: "ghost" }}
-            href={{
-              search: { edit: true },
-            }}
-          >
-            <Pencil className="size-3.5" /> Edit
-          </Link>
-        )}
-      </Authenticated>
-    </div>
-  );
-}
 
 interface QuestPageHeaderProps extends Omit<PageHeaderProps, "title"> {
   quest?: Doc<"quests"> | null;
@@ -83,11 +25,9 @@ export function QuestPageHeader({
   quest,
   userQuest,
   editable,
-  children,
 }: QuestPageHeaderProps) {
-  const [isEditingTitle, setIsEditingTitle] = useState(false);
-  const isMobile = useIsMobile();
   const navigate = useNavigate();
+  const [isEditingTitle, setIsEditingTitle] = useState(false);
   const changeStatus = useMutation(api.userQuests.setStatus);
   const deleteForever = useMutation(api.userQuests.deleteForever);
 
@@ -104,51 +44,43 @@ export function QuestPageHeader({
     if (quest) changeStatus({ questId: quest._id, status });
   };
 
+  const editTitleButton = (
+    <>
+      <TooltipTrigger>
+        <Button
+          variant="icon"
+          size="small"
+          icon={Pencil}
+          onPress={() => setIsEditingTitle(true)}
+        />
+        <Tooltip placement="right">Edit title</Tooltip>
+      </TooltipTrigger>
+      <EditQuestTitleModal
+        quest={quest}
+        open={isEditingTitle}
+        onOpenChange={setIsEditingTitle}
+      />
+    </>
+  );
+
   return (
-    <div className="pt-6 mb-6 flex flex-col gap-2 items-start w-full border-b border-gray-dim">
-      <div className="flex gap-1 items-center">
-        <QuestCategory quest={quest} editable={editable} />
-        <QuestJurisdiction quest={quest} editable={editable} />
-        <QuestCosts quest={quest} editable={editable} />
-        <QuestTimeRequired quest={quest} editable={editable} />
-      </div>
-      <div className="flex gap-2 items-center">
-        {isMobile && (
-          <Link button={{ variant: "icon" }} href={{ to: "/" }}>
-            <ArrowLeft className="size-6" />
-          </Link>
+    <>
+      <PageHeader
+        title={quest.title}
+        badge={editable ? editTitleButton : undefined}
+        mobileBackLink={{ to: "/" }}
+      >
+        {quest && userQuest && !editable && (
+          <StatusSelect
+            status={userQuest.status as Status}
+            onChange={handleStatusChange}
+            onRemove={() => handleRemoveQuest(quest._id, quest.title)}
+          />
         )}
-        <Heading className="text-4xl font-medium">{quest?.title}</Heading>
-        {editable && (
-          <>
-            <TooltipTrigger>
-              <Button
-                variant="icon"
-                size="small"
-                icon={Pencil}
-                onPress={() => setIsEditingTitle(true)}
-              />
-              <Tooltip placement="bottom">Edit title</Tooltip>
-            </TooltipTrigger>
-            <EditQuestTitleModal
-              quest={quest}
-              open={isEditingTitle}
-              onOpenChange={setIsEditingTitle}
-            />
-          </>
-        )}
-        <div className="flex gap-2 items-center -mb-1">
-          {quest && userQuest && !editable && (
-            <StatusSelect
-              status={userQuest.status as Status}
-              onChange={handleStatusChange}
-              onRemove={() => handleRemoveQuest(quest._id, quest.title)}
-            />
-          )}
-          {children}
-        </div>
-      </div>
-      <QuestPageToolbar quest={quest} editable={editable} />
-    </div>
+        {}
+      </PageHeader>
+      <QuestPageBadges quest={quest} editable={editable} />
+      <QuestPageToolbar quest={quest} isEditing={editable} />
+    </>
   );
 }
