+++ src/components/quests/badges/QuestJurisdictionBadge/QuestJurisdictionBadge.tsx
@@ -1,15 +1,20 @@
 import {
+  Badge,
+  BadgeButton,
   Button,
   Form,
   Modal,
   ModalHeader,
   Select,
   SelectItem,
+  Tooltip,
+  TooltipTrigger,
 } from "@/components/common";
 import { api } from "@convex/_generated/api";
 import type { Doc } from "@convex/_generated/dataModel";
 import { JURISDICTIONS, type Jurisdiction } from "@convex/constants";
 import { useMutation } from "convex/react";
+import { Pencil } from "lucide-react";
 import { useState } from "react";
 import { toast } from "sonner";
 
@@ -86,5 +91,53 @@ export const EditQuestJurisdictionModal = ({
         </div>
       </Form>
     </Modal>
+  );
+};
+
+type QuestJurisdictionBadgeProps = {
+  quest?: Doc<"quests"> | null;
+  editable?: boolean;
+};
+
+export const QuestJurisdictionBadge = ({
+  quest,
+  editable = false,
+}: QuestJurisdictionBadgeProps) => {
+  const [isEditing, setIsEditing] = useState(false);
+
+  if (!quest) return null;
+
+  function getJurisdictionLabel(quest: Doc<"quests">) {
+    const { jurisdiction, category } = quest;
+    if (category === "passport" || category === "socialSecurity") {
+      return "United States";
+    }
+
+    if (jurisdiction) {
+      return JURISDICTIONS[jurisdiction as keyof typeof JURISDICTIONS];
+    }
+  }
+
+  return (
+    <Badge>
+      {getJurisdictionLabel(quest)}
+      {editable && (
+        <>
+          <TooltipTrigger>
+            <BadgeButton
+              icon={Pencil}
+              onPress={() => setIsEditing(true)}
+              label="Edit jurisdiction"
+            />
+            <Tooltip>Edit state</Tooltip>
+          </TooltipTrigger>
+          <EditQuestJurisdictionModal
+            quest={quest}
+            open={isEditing}
+            onOpenChange={setIsEditing}
+          />
+        </>
+      )}
+    </Badge>
   );
 };
