+++ src/components/quests/badges/index.ts
@@ -1,0 +1,4 @@
+export * from "./QuestCategoryBadge/QuestCategoryBadge";
+export * from "./QuestCostsBadge/QuestCostsBadge";
+export * from "./QuestJurisdictionBadge/QuestJurisdictionBadge";
+export * from "./QuestTimeBadge/QuestTimeBadge";
