+++ src/components/quests/index.ts
@@ -1,19 +1,14 @@
-export * from "./EditQuestCategoryModal/EditQuestCategoryModal";
-export * from "./EditQuestCostsModal/EditQuestCostsModal";
-export * from "./EditQuestTimeRequiredModal/EditQuestTimeRequiredModal";
 export * from "./EditQuestTitleModal/EditQuestTitleModal";
 export * from "./JurisdictionInterstitial/JurisdictionInterstitial";
-export * from "./QuestCategory/QuestCategory";
+export * from "./QuestCallToAction/QuestCallToAction";
 export * from "./QuestContent/QuestContent";
-export * from "./QuestCosts/QuestCosts";
 export * from "./QuestFaqs/QuestFaqs";
-export * from "./QuestJurisdiction/QuestJurisdiction";
+export * from "./QuestPageBadges/QuestPageBadges";
 export * from "./QuestPageHeader/QuestPageHeader";
+export * from "./QuestPageToolbar/QuestPageToolbar";
 export * from "./QuestSection/QuestSection";
 export * from "./QuestsNav/QuestsNav";
-export * from "./QuestStatusFooter/QuestStatusFooter";
 export * from "./QuestSteps/QuestSteps";
-export * from "./QuestTimeRequired/QuestTimeRequired";
 export * from "./ReadingScore/ReadingScore";
 export * from "./StatusSelect/StatusSelect";
 export * from "./Steps/Steps";
