+++ src/routes/_authenticated/_home/quests.$questSlug.index.tsx
@@ -1,14 +1,16 @@
 import { Empty } from "@/components/common";
 import {
+  QuestCallToAction,
   QuestFaqs,
   QuestPageHeader,
-  QuestStatusFooter,
   QuestSteps,
 } from "@/components/quests";
+import { useIsMobile } from "@/utils/useIsMobile";
 import { api } from "@convex/_generated/api";
 import { createFileRoute } from "@tanstack/react-router";
 import { useQuery } from "convex/react";
 import { Milestone } from "lucide-react";
+import { tv } from "tailwind-variants";
 
 type QuestSearch = {
   edit?: true | undefined;
@@ -28,6 +30,7 @@ export const Route = createFileRoute(
 function QuestDetailRoute() {
   const { edit: isEditing } = Route.useSearch();
   const { questSlug } = Route.useParams();
+  const isMobile = useIsMobile();
 
   const questData = useQuery(api.quests.getWithUserQuest, {
     slug: questSlug,
@@ -51,20 +54,28 @@ function QuestDetailRoute() {
   const quest = questData.quest;
   const userQuest = questData.userQuest;
 
+  const containerStyles = tv({
+    base: "flex flex-col",
+    variants: {
+      isMobile: {
+        true: "min-h-screen-minus-mobile",
+        false: "min-h-dvh",
+      },
+    },
+  });
+
   return (
-    <>
+    <div className={containerStyles({ isMobile })}>
       <QuestPageHeader
         quest={quest}
         userQuest={userQuest}
         editable={isEditing}
       />
-      <div className="flex flex-col gap-6 pb-12">
+      <div className="flex flex-1 flex-col gap-6 app-padding">
         <QuestSteps quest={quest} editable={isEditing} />
         <QuestFaqs quest={quest} editable={isEditing} />
-        {!isEditing && (
-          <QuestStatusFooter quest={quest} userQuest={userQuest} />
-        )}
       </div>
-    </>
+      {!isEditing && <QuestCallToAction quest={quest} userQuest={userQuest} />}
+    </div>
   );
 }
