+++ src/routes/_authenticated/settings/account.tsx
@@ -28,7 +28,7 @@ function SettingsAccountRoute() {
       ) : user === null ? (
         "User not found, please reload"
       ) : (
-        <div className="pb-8">
+        <div className="app-padding pb-8">
           <SettingsGroup title="Personal Information">
             <EditNameSetting user={user} />
             <EditEmailSetting user={user} />
